requirements.txt
Django==4.2
flake8==6.1.0
psycopg2>=2.8.6,<2.9
Pillow>=8.2.0,<8.3.0
django-simple-captcha==0.6.0
django-mptt>=0.12.0